Break apart keyboard-navigable-menu into modules
