+++ utilities/keyboard-navigable-menu.js
@@ -1,43 +1,19 @@
 /* Copyright (c) 2015-present, salesforce.com, inc. All rights reserved */
 /* Licensed under BSD 3-Clause - see LICENSE.txt or git.io/sfdc-license */
 
-// # Keyboard Navigable Trait
-
-// ## Dependencies
-
-// ### React
+// # Keyboard Navigable mixin (deprecated. DO NOT USE)
 import ReactDOM from 'react-dom';
-
-// ### escapeRegExp
-import escapeRegExp from 'lodash.escaperegexp';
 
 // ### isFunction
 import isFunction from 'lodash.isfunction';
 
 // ### Event Helpers
-import KEYS from './key-code';
+import KeyBuffer from './key-buffer';
+import keyboardNavigate from './keyboard-navigate';
 
 /* eslint-disable react/no-find-dom-node */
 
 const noop = () => {};
-
-export function KeyBuffer() {
-	this.buffer = '';
-
-	return (key) => {
-		if (this.timeout) {
-			clearTimeout(this.timeout);
-			this.timeout = undefined;
-		}
-
-		this.timeout = setTimeout(() => {
-			this.buffer = '';
-		}, 400);
-
-		this.buffer = this.buffer + key;
-		return this.buffer;
-	};
-}
 
 export function itemIsSelectable(item) {
 	return item.type !== 'header' && item.type !== 'divider' && !item.disabled;
@@ -62,98 +38,6 @@ export function getNavigableItems(items) {
 	}
 
 	return navigableItems;
-}
-
-export function keyboardNavigate({
-	componentContext,
-	currentFocusedIndex,
-	isOpen,
-	event,
-	key,
-	keyCode,
-	navigableItems,
-	onFocus,
-	onSelect,
-	target,
-	toggleOpen,
-}) {
-	const indexes = navigableItems.indexes;
-	const lastIndex = indexes.length - 1;
-	let focusedIndex;
-	let ch = key || String.fromCharCode(keyCode);
-
-	if (/^[ -~]$/.test(ch)) {
-		ch = ch.toLowerCase();
-	} else {
-		ch = null;
-	}
-
-	const openMenuKeys =
-		keyCode === KEYS.ENTER || keyCode === KEYS.SPACE || keyCode === KEYS.UP;
-
-	if (keyCode === KEYS.ESCAPE) {
-		if (isOpen) toggleOpen();
-	} else if (!isOpen) {
-		focusedIndex = indexes[0];
-		if (openMenuKeys || ch) {
-			toggleOpen();
-		}
-		if (
-			openMenuKeys &&
-			componentContext.trigger &&
-			ReactDOM.findDOMNode(componentContext.trigger) === target
-		) {
-			// eslint-disable-line react/no-find-dom-node
-			componentContext.handleClick(event);
-		}
-	} else if (keyCode === KEYS.ENTER || keyCode === KEYS.SPACE) {
-		onSelect(currentFocusedIndex);
-	} else {
-		const navigableIndex = indexes.indexOf(currentFocusedIndex);
-
-		if (keyCode === KEYS.DOWN) {
-			if (navigableIndex < lastIndex) {
-				const newNavigableIndex = navigableIndex + 1;
-				focusedIndex = indexes[newNavigableIndex];
-			} else {
-				focusedIndex = indexes[0];
-			}
-		} else if (keyCode === KEYS.UP) {
-			if (navigableIndex > 0) {
-				const newNavigableIndex = navigableIndex - 1;
-				focusedIndex = indexes[newNavigableIndex];
-			} else {
-				focusedIndex = indexes[lastIndex];
-			}
-		} else if (ch) {
-			// Combine subsequent keypresses
-			const pattern = navigableItems.keyBuffer(ch);
-			let consecutive = 0;
-
-			// Support for navigating to the next option of the same letter with repeated presses of the same key
-			if (
-				pattern.length > 1 &&
-				new RegExp(`^[${escapeRegExp(ch)}]+$`).test(pattern)
-			) {
-				consecutive = pattern.length;
-			}
-
-			navigableItems.forEach((item) => {
-				if (
-					(focusedIndex === undefined &&
-						item.text.substr(0, pattern.length) === pattern) ||
-					(consecutive > 0 && item.text.substr(0, 1) === ch)
-				) {
-					consecutive -= 1;
-					focusedIndex = item.index;
-				}
-			});
-		}
-	}
-
-	onFocus(focusedIndex);
-
-	return focusedIndex;
 }
 
 function getMenu(componentRef) {
